Created initial form, need check box
