+++ aunt-leahs-app/src/App.js
@@ -1,9 +1,8 @@
 import React from 'react';
 
 import './App.css';
-import HomePage from './pages/homepage/homepage';
-import NewVolunteer from './pages/new-volunteer/new-volunteer';
 import HomePage from './pages/homepage';
+import NewVolunteer from './pages/new-volunteer';
 import './styles.css'
 
 function App() {
